# padariaSwagger/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "PADOCA",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/funcionario": {
      "get": {
        "tags": [
          "Funcionário"
        ],
        "description": "Endpoint (procedure) para obter todos os funcionarios.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Funcionário"
        ],
        "description": "Endpoint para cadastrar um usuário.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Objeto para criação de Funcionário",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nivelAcesso": {
                  "type": "string",
                  "example": "master"
                },
                "nome": {
                  "type": "string",
                  "example": "Izauir Guilherme Bernardo dos Santos"
                },
                "sexo": {
                  "type": "string",
                  "example": "M"
                },
                "cargo": {
                  "type": "string",
                  "example": "Programador"
                },
                "nascimento": {
                  "type": "string",
                  "example": "1999-09-18"
                },
                "telefone": {
                  "type": "string",
                  "example": "(11)97369-9232"
                },
                "email": {
                  "type": "string",
                  "example": "izauirguilherme@hotmail.com"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/funcionario/{id}": {
      "put": {
        "tags": [
          "Funcionário"
        ],
        "description": "Endpoint para atualizar o Funcionário.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "User ID."
          },
          {
            "name": "body",
            "in": "body",
            "description": "Objeto para criação de Funcionário",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nivelAcesso": {
                  "type": "string",
                  "example": "master"
                },
                "nome": {
                  "type": "string",
                  "example": "Izauir Guilherme Bernardo dos Santos"
                },
                "sexo": {
                  "type": "string",
                  "example": "M"
                },
                "cargo": {
                  "type": "string",
                  "example": "Programador"
                },
                "nascimento": {
                  "type": "string",
                  "example": "1999-09-18"
                },
                "telefone": {
                  "type": "string",
                  "example": "(11)97369-9232"
                },
                "email": {
                  "type": "string",
                  "example": "izauirguilherme@hotmail.com"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "patch": {
        "tags": [
          "Funcionário"
        ],
        "description": "Endpoint para atualizar apenas o cargo do Funcionário.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Objeto com o cargo desejado",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "cargo": {
                  "type": "string",
                  "example": "Streamer"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/estoque": {
      "get": {
        "tags": [
          "Estoque"
        ],
        "description": "Endpoint (procedure) para obter todos os produtos no estoques da padaria.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Estoque"
        ],
        "description": "Endpoint para cadastrar um produto no estoque.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Objeto para criação de Estoque",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "idProduto": {
                  "type": "number",
                  "example": 1
                },
                "quantidade": {
                  "type": "number",
                  "example": 100
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/estoque/{id}": {
      "get": {
        "tags": [
          "Estoque"
        ],
        "description": "Endpoint trazer apenas um produto por id no Estoque.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "in": "params",
          "required": true
        }
      },
      "put": {
        "tags": [
          "Estoque"
        ],
        "description": "Endpoint para atualizar um produto do Estoque.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Objeto para criação de Estoque",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "idProduto": {
                  "type": "number",
                  "example": 1
                },
                "quantidade": {
                  "type": "number",
                  "example": 100
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
  }
}